+++ src/main/java/nl/talsmasoftware/umldoclet/javadoc/dependencies/PackageDependenciesElementScanner.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.talsmasoftware.umldoclet.javadoc.packagedependencies;
+package nl.talsmasoftware.umldoclet.javadoc.dependencies;
 
 import nl.talsmasoftware.umldoclet.uml.Namespace;
 import nl.talsmasoftware.umldoclet.uml.Reference;
